better naming and comments added
