+++ poetrysite-shiny/README.md
@@ -1,14 +1,10 @@
 # poetrysite-shiny
 
-This template should help get you started developing with Vue 3 in Vite.
+This Single Page Application allows you to browse a collection of poems and poets from the PoetryDB API
 
 ## Recommended IDE Setup
 
 [VSCode](https://code.visualstudio.com/) + [Volar](https://marketplace.visualstudio.com/items?itemName=Vue.volar) (and disable Vetur) + [TypeScript Vue Plugin (Volar)](https://marketplace.visualstudio.com/items?itemName=Vue.vscode-typescript-vue-plugin).
-
-## Customize configuration
-
-See [Vite Configuration Reference](https://vitejs.dev/config/).
 
 ## Project Setup
