+++ poetrysite-shiny/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="https://use.fontawesome.com/releases/v5.15.1/css/all.css" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700,900&display=swap" rel="stylesheet" />
-    <title>Vite App</title>
+    <title>Poetry Fetch</title>
   </head>
   <body>
     <div id="app"></div>
