+++ poetrysite-shiny/src/App.vue
@@ -1,16 +1,24 @@
 <script setup>
-import HelloWorld from './components/HelloWorld.vue'
+//This is the main app vue file
+//import components
+import Authors from './components/Authors.vue'
 import NavBar from './components/NavBar.vue';
+import Random from './components/Random.vue';
 </script>
 
 <template>
   <header>
-<NavBar></NavBar>
+    <NavBar></NavBar>
   </header>
 
   <main>
-      <div class="wrapper">
-      <HelloWorld/>
+    <div class="wrapper">
+      <Authors />
+    </div>
+
+
+    <div class="wrapper">
+      <Random></Random>
     </div>
   </main>
 </template>
