--- poetrysite-shiny/src/components/HelloWorld.vue
@@ -1,38 +1,0 @@
-<script>
-import axios from 'axios';
-
-export default {
-  data: () => ({
-    lines: ['Loading...'],
-    author: "Author",
-    title: "Title"
-  }),
-  created() {
-    // fetch on init
-    this.fetchData()
-  },
-  methods: {
-    //an asynchronous function to get data from the api and assign it to *
-    async fetchData() {
-      const response = await axios.get('https://poetrydb.org/random');
-      this.title = (response.data[0].title);
-      this.lines = (response.data[0].lines);
-      this.author = response.data[0].author;
-      
-
-    }
-  }
-};
-</script>
-
-
-
-<template>
-  <h3>{{ header }}</h3>
-
-  <template v-for="author in authors">
-    <ul>
-      <label>{{ author }}</label>
-    </ul>
-  </template>
-</template>
