+++ poetrysite-shiny/src/main.js
@@ -2,5 +2,8 @@ import { createApp } from 'vue'
 import App from './App.vue'
 import 'mdb-vue-ui-kit/css/mdb.min.css';
 
+//import and use router
 
+
+//mount app
 createApp(App).mount('#app')
